+++ src/LIAnsureProtect.Worker/Program.cs
@@ -1,6 +1,10 @@
+using LIAnsureProtect.Application;
+using LIAnsureProtect.Infrastructure;
 using LIAnsureProtect.Worker;
 
 var builder = Host.CreateApplicationBuilder(args);
+builder.Services.AddApplication();
+builder.Services.AddInfrastructure();
 builder.Services.AddHostedService<Worker>();
 
 var host = builder.Build();